+++ lib/src/lcu_http_client.dart
@@ -8,6 +8,43 @@ import 'package:gangplank/src/lcu_watcher.dart';
 import 'package:gangplank/src/logging.dart';
 import 'package:gangplank/src/storage.dart';
 
+enum LCUGetRouteToCacheMatchType {
+  equals,
+  contains,
+  startsWith,
+  endsWith,
+}
+
+class LCUGetRouteToCache {
+  /// The route that must be cached.
+  final String route;
+
+  /// The lifetime of this specific object in the cache.
+  /// 
+  /// If not set, [cacheExpiration] defaults to the global duration.
+  final Duration? cacheExpiration;
+
+  /// The [LCUGetRouteToCacheMatchType] determines on how the route will be matched.
+  final LCUGetRouteToCacheMatchType matchType;
+
+  LCUGetRouteToCache({
+    required this.route,
+    this.cacheExpiration,
+    this.matchType = LCUGetRouteToCacheMatchType.contains,
+  }): assert(route.split('*').length <= 2, 'ONLY ONE WILDCARD IS ALLOWED.');
+
+  Map<String, dynamic> toJson() => {
+    'route': route,
+    'cacheExpiration': cacheExpiration?.inSeconds,
+    'matchType': matchType.name,
+  };
+
+  @override
+  String toString() {
+    return jsonEncode(toJson());
+  }
+}
+
 class LCUHttpClientConfig {
   /// Disables the logging for the [LCUHttpClient].
   /// 
@@ -26,17 +63,29 @@ class LCUHttpClientConfig {
   /// This builds up alot of requests to the summoner endpoint.
   /// Since the summoner data won't change frequently you can supply the endpoint route here to cache it.
   ///
-  /// The application will then take the data out of cache instead of requesting the data from LCU.
+  /// The application will then grab the data out of cache instead of requesting the data from LCU.
   /// 
-  /// The routes are matched via contains and not equal which means if you supply the endpoint `/lol-summoner/v1/summoners` all routes will be cached that include that string.
+  /// You can decide your on own how the routes shall be cached based on matchType.
+  /// 
+  /// [LCUGetRouteToCacheMatchType.equals] will only cache the route if it the requested route matches 100%.
+  /// If you supply `/lol-summoner/v1/summoners` the requested route `/lol-summoner/v1/summoners?name=test` won't be cached, but `/lol-summoner/v1/summoners` will.
+  /// 
+  /// [LCUGetRouteToCacheMatchType.contains] will only cache the route if it the requested route contains given value.
+  /// If you supply `/lol-summoner/v1/summoners` the requested route `/lol-summoner/v1/summoners?name=test` will be cached as well as `/lol-summoner/v1/summoners`.
+  /// 
+  /// [LCUGetRouteToCacheMatchType.startsWith] will only cache the route if it the requested route starts with given value.
+  /// If you supply `/lol-summoner/v1/` the requested route `/lol-summoner/v1/summoners?name=test` will be cached.
+  /// 
+  /// [LCUGetRouteToCacheMatchType.endsWith] will only cache the route if it the requested route ends with given value.
+  /// If you supply `/v1/summoners` the requested route `/lol-summoner/v1/summoners?name=test` won't be cached, but `/lol-summoner/v1/summoners` will.
+  /// 
   /// In this case `/lol-summoner/v1/summoners?name=test` e.g. will be cached.
-  final List<String> getRoutesToCache;
+  final List<LCUGetRouteToCache> getRoutesToCache;
 
   /// The lifetime of an object in the cache.
   /// 
   /// [cacheExpiration] defaults to 30 minutes.
   final Duration cacheExpiration;
-
 
   LCUHttpClientConfig({ 
     this.disableLogging = false, 
@@ -79,6 +128,8 @@ class LCUHttpClient {
       service: 'LCU-HTTP-CLIENT',
       storage: _storage,
     );
+
+    cache.clear();
   }
 
   /// Fires a GET request against the League client.
@@ -104,10 +155,16 @@ class LCUHttpClient {
       params: params,
     );
 
-    if (_config.getRoutesToCache.any((e) => endpoint.toLowerCase().contains(e.toLowerCase()))) {
-      // ROUTE SHOULD BE CACHED
+    final LCUGetRouteToCache? foundEntry = _mustCacheEndpoint(endpoint);
 
-      await cache.write(endpoint, result, _config.cacheExpiration.inSeconds);
+    if (foundEntry != null) {
+      // ENTRY FOUND!
+      // ROUTE SHOULD BE CACHED
+      // IF CACHE EXPIRATION SET ON FOUND ENTRY USE IT, OTHERWISE GLOBAL CACHE EXPIRATION
+
+      await cache.write(endpoint, result, foundEntry.cacheExpiration != null ? foundEntry.cacheExpiration!.inSeconds : _config.cacheExpiration.inSeconds);
+
+      if (!_config.disableLogging) _logger.log('CACHED: $endpoint FOR ${foundEntry.cacheExpiration != null ? foundEntry.cacheExpiration!.inSeconds : _config.cacheExpiration.inSeconds} SECONDS');
     }
 
     return result;
@@ -296,5 +353,41 @@ class LCUHttpClient {
         httpStatus: 400,
       );
     }
+  }
+
+  /// Checks if [endpoint] must be cached
+  /// 
+  /// Checks by matchtype and returns [LCUGetRouteToCache] if a route matches or returns `null` if it doesn't.
+  LCUGetRouteToCache? _mustCacheEndpoint(String endpoint) {
+    String lcEndpoint = endpoint.toLowerCase();
+
+    for(LCUGetRouteToCache getRouteToCache in _config.getRoutesToCache) {
+      switch(getRouteToCache.matchType) {
+        case LCUGetRouteToCacheMatchType.equals:
+          if (lcEndpoint == getRouteToCache.route.toLowerCase()) return getRouteToCache;
+          return null;
+        case LCUGetRouteToCacheMatchType.contains:
+          if (lcEndpoint.contains(getRouteToCache.route.toLowerCase())) return getRouteToCache;
+          return null;
+        case LCUGetRouteToCacheMatchType.startsWith:
+          if (lcEndpoint.startsWith(getRouteToCache.route.toLowerCase())) return getRouteToCache;
+          return null;
+        case LCUGetRouteToCacheMatchType.endsWith:
+          if (lcEndpoint.endsWith(getRouteToCache.route.toLowerCase())) return getRouteToCache;
+          return null;
+        default:
+          return null;
+      }
+    }
+
+    return null;
+  }
+
+  void clearCache() {
+    cache.clear();
+  }
+
+  void removeKeyFromCache(String url) {
+    cache.destroy(url);
   }
 }